# web-ui-specs/02-api-specification.yaml
openapi: 3.0.0
info:
  title: OpenFederation PDS Web Interface API
  version: 1.0.0
  description: |
    API specification for OpenFederation Personal Data Server Web Management Interface.
    Built on AT Protocol (XRPC) for decentralized identity and community management.
    
    **Key Features:**
    - AT Protocol native endpoints (XRPC)
    - DID-based identity (did:plc and did:web)
    - Community sovereignty and portability
    - JWT-based session management
    
    **Base URL:** All endpoints are prefixed with `/xrpc/` following AT Protocol conventions.
  
  contact:
    name: OpenFederation
    url: https://openfederation.net
  
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://pds.openfederation.net
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /xrpc/net.openfederation.invite.create:
    post:
      summary: Create invite code
      description: Generate a new invite code (admin only)
      tags: [Admin, Invites]
      operationId: createInvite
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                maxUses:
                  type: integer
                  minimum: 1
                  default: 1
                  description: Maximum number of times code can be used
                expiresIn:
                  type: integer
                  minimum: 3600
                  description: Expiration time in seconds (optional)
                  example: 86400
      responses:
        '201':
          description: Invite code created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteCode'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== COMMUNITY MANAGEMENT ====================
  /xrpc/net.openfederation.community.create:
    post:
      summary: Create new community
      description: |
        Create a new community with DID method choice (did:plc or did:web).
        User becomes the community owner.
        Account must be approved to create communities.
      tags: [Communities]
      operationId: createCommunity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [handle, didMethod]
              properties:
                handle:
                  type: string
                  pattern: '^[a-z0-9-]+$'
                  minLength: 3
                  maxLength: 255
                  description: Community handle (unique identifier)
                  example: "my-community"
                didMethod:
                  type: string
                  enum: [plc, web]
                  description: DID method to use
                  example: "plc"
                domain:
                  type: string
                  format: hostname
                  description: Domain for did:web (required if didMethod is 'web')
                  example: "example.com"
                displayName:
                  type: string
                  maxLength: 255
                  description: Community display name (optional)
                  example: "My Cool Community"
                description:
                  type: string
                  description: Community description (optional)
                  example: "A community for cool people"
      responses:
        '201':
          description: Community created successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CommunityPLC'
                  - $ref: '#/components/schemas/CommunityWeb'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account not approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Handle already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /xrpc/com.atproto.repo.getRecord:
    get:
      summary: Get AT Protocol record
      description: |
        Fetch a single record from a repository.
        Used to get community profiles and settings.
      tags: [Communities]
      operationId: getRecord
      parameters:
        - name: repo
          in: query
          required: true
          schema:
            type: string
          description: DID of the repository
          example: "did:plc:xyz123"
        - name: collection
          in: query
          required: true
          schema:
            type: string
          description: Collection name
          example: "net.openfederation.community.profile"
        - name: rkey
          in: query
          required: true
          schema:
            type: string
          description: Record key
          example: "self"
      responses:
        '200':
          description: Record retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ATProtocolRecord'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

# ═══════════════════════════════════════════════════════════════
# REUSABLE COMPONENTS
# ═══════════════════════════════════════════════════════════════
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT token (15 min TTL)
    
    refreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh JWT token (30 days TTL)

  # ==================== SCHEMAS ====================
  schemas:
...
    InviteCode:
      type: object
      properties:
        code:
          type: string
          description: Generated invite code
          example: "ABC123XYZ"
        maxUses:
          type: integer
          example: 1
        usesCount:
          type: integer
          example: 0
        expiresAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-01-02T00:00:00Z"
        createdBy:
          type: string
          format: uuid
          description: Admin user ID who created this invite
    CommunityPLC:
      type: object
      description: Community created with did:plc method
      properties:
        did:
          type: string
          pattern: '^did:plc:'
          example: "did:plc:xyz123"
        handle:
          type: string
          example: "my-community"
        didMethod:
          type: string
          enum: [plc]
        primaryRotationKey:
          type: string
          description: Base64-encoded rotation key (BACKUP IMMEDIATELY)
          example: "base64-encoded-key..."
        message:
          type: string
          example: "IMPORTANT: Please back up your primaryRotationKey immediately. This is the only time it will be shown."
    CommunityWeb:
      type: object
      description: Community created with did:web method
      properties:
        did:
          type: string
          pattern: '^did:web:'
          example: "did:web:example.com"
        handle:
          type: string
          example: "my-community"
        didMethod:
          type: string
          enum: [web]
        didDocument:
          type: object
          description: DID document to host
        instructions:
          type: string
          example: "Host this JSON document at https://example.com/.well-known/did.json"
    ATProtocolRecord:
      type: object
      properties:
        uri:
          type: string
          description: AT Protocol URI
          example: "at://did:plc:xyz123/net.openfederation.community.profile/self"
        cid:
          type: string
          description: Content identifier
          example: "bafyreiabc123..."
        value:
          type: object
          description: Record value (schema depends on collection)
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
          example: "InvalidRequest"
        message:
          type: string
          description: Human-readable error message
          example: "Missing required field: handle"

  # ==================== RESPONSES ====================
  responses:
    InvalidRequest:
      description: Invalid request (malformed, missing fields)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "InvalidRequest"
            message: "Missing required field: handle"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "AuthenticationRequired"
            message: "Missing or invalid authentication token"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden"
            message: "Insufficient permissions for this operation"
...